feat(dedicated): refactoring diagnostic download
 ref: MANAGER-16172
