+++ packages/manager/modules/cloud-connect/src/details/diagnostics/diagnostics.controller.js
@@ -1,0 +1,15 @@
+export default class DiagnosticsResultCtrl {
+  /* @ngInject */
+  constructor(cloudConnectService, cloudConnectDiagnosticsService) {
+    this.cloudConnectService = cloudConnectService;
+    this.cloudConnectDiagnosticsService = cloudConnectDiagnosticsService;
+  }
+
+  downloadResult(diagnosticId) {
+    this.cloudConnectService
+      .getDiagnostic(this.cloudConnect.id, diagnosticId)
+      .then((diagnostic) => {
+        this.cloudConnectDiagnosticsService.download(diagnostic);
+      });
+  }
+}
