+++ packages/manager/modules/cloud-connect/src/details/diagnostics/diagnostics.module.js
@@ -2,6 +2,8 @@ import angular from 'angular';
 
 import component from './diagnostics.component';
 import routing from './diagnostics.routing';
+import service from './diagnostics.service';
+
 import ovhCloudConnectDiagnosticsResult from './result';
 
 const moduleName = 'ovhCloudConnectDiagnosticsDetails';
@@ -9,6 +11,7 @@ const moduleName = 'ovhCloudConnectDiagnosticsDetails';
 angular
   .module(moduleName, [ovhCloudConnectDiagnosticsResult])
   .config(routing)
+  .service('cloudConnectDiagnosticsService', service)
   .component('cloudConnectDiagnosticsDetails', component)
   .run(/* @ngTranslationsInject:json ./translations */);
