+++ packages/manager/modules/cloud-connect/src/details/diagnostics/result/template.html
@@ -56,7 +56,7 @@
             </button>
             <button
                 class="oui-button oui-button_s oui-button_ghost"
-                data-ng-click="$ctrl.constructor.download($ctrl.diagnostic)"
+                data-ng-click="$ctrl.download($ctrl.diagnostic)"
             >
                 <span class="fa fa-download"></span>
             </button>
