+++ packages/manager/modules/cloud-connect/src/details/overview/template.html
@@ -430,36 +430,46 @@
                             data-term="{{ ::'cloud_connect_pop_diagnostic' | translate }}"
                         >
                             <oui-tile-description>
-                                <a
-                                    class="oui-link_icon"
-                                    data-ng-click="$ctrl.goToDiagnosticPage()"
-                                >
-                                    <span
-                                        data-translate="cloud_connect_pop_diagnostic_link"
-                                    ></span>
-                                    <span
-                                        class="oui-icon oui-icon-arrow-right align-middle d-inline-block"
-                                        aria-hidden="true"
-                                    ></span>
-                                </a>
+                                <div class="row d-lg-flex">
+                                    <div class="col-xm-12 col-md-9">
+                                        <a
+                                            class="oui-link_icon"
+                                            data-ng-href="$ctrl.goToDiagnosticPage()"
+                                            data-on-click="$ctrl.goToDiagnosticPage()"
+                                        >
+                                            <span
+                                                data-translate="cloud_connect_pop_diagnostic_link"
+                                            ></span>
+                                            <span
+                                                aria-hidden="true"
+                                                class="oui-icon oui-icon-arrow-right align-middle d-inline-block"
+                                            ></span>
+                                        </a>
+                                    </div>
+                                    <div class="col-xm-12 col-md-3">
+                                        <oui-action-menu
+                                            data-compact
+                                            data-placement="end"
+                                        >
+                                            <oui-action-menu-item
+                                                data-ng-if="$pop.isL3Type()"
+                                                data-on-click="$ctrl.goToCheckBGPPeeringPage({popConfigId: $pop.id , dcConfigId: $ctrl.datacenters.id })"
+                                                ><span
+                                                    data-translate="cloud_connect_pop_diagnostic_check"
+                                                ></span
+                                            ></oui-action-menu-item>
+                                            <oui-action-menu-item
+                                                data-ng-if="$pop.isL2Type()"
+                                                data-on-click="$ctrl.getMacList($pop)"
+                                                data-disabled="$ctrl.getMacLoading"
+                                                ><span
+                                                    data-translate="cloud_connect_pop_get_mac_list"
+                                                ></span
+                                            ></oui-action-menu-item>
+                                        </oui-action-menu>
+                                    </div>
+                                </div>
                             </oui-tile-description>
-                            <oui-action-menu data-compact data-placement="end">
-                                <oui-action-menu-item
-                                    data-ng-if="$pop.isL3Type()"
-                                    data-on-click="$ctrl.goToCheckBGPPeeringPage({popConfigId: $pop.id , dcConfigId: $ctrl.datacenters.id })"
-                                    ><span
-                                        data-translate="cloud_connect_pop_diagnostic_check"
-                                    ></span
-                                ></oui-action-menu-item>
-                                <oui-action-menu-item
-                                    data-ng-if="$pop.isL2Type()"
-                                    data-on-click="$ctrl.getMacList($pop)"
-                                    data-disabled="$ctrl.getMacLoading"
-                                    ><span
-                                        data-translate="cloud_connect_pop_get_mac_list"
-                                    ></span
-                                ></oui-action-menu-item>
-                            </oui-action-menu>
                         </oui-tile-definition>
                     </div>
                 </div>
