+++ packages/manager/modules/cloud-connect/src/details/overview/translations/Messages_fr_FR.json
@@ -38,7 +38,7 @@
   "cloud_connect_network": "Sous-réseau",
   "cloud_connect_network_address_placeholder": "Saisissez Sous-réseau en /30",
   "cloud_connect_pop_diagnostic": "POP Diagnostic",
-  "cloud_connect_pop_diagnostic_link": "Voir les résultats de mes diagnostic",
+  "cloud_connect_pop_diagnostic_link": "Voir les résultats </br> de mes diagnostic",
   "cloud_connect_pop_diagnostic_check": "Test BGP Peering",
   "cloud_connect_pop_get_mac_list": "Voir la liste de mes adresses MAC",
   "cloud_connect_provider": "Provider",
